新增BoxSelectionTool工具的引用，优化框选工具的实现，移除不必要的渲染类和代理逻辑。
